fix: move validators to dedicated directory
