+++ internal/validators/email_validator.go
@@ -1,4 +1,4 @@
-package utils
+package validators
 
 import (
 	"errors"
